refactor: file select and uploadimage are in the one composable
