+++ e2e/vue.spec.ts
@@ -1,8 +1,5 @@
-import { test, expect } from '@playwright/test';
+import { test } from '@playwright/test';
 
-// See here how to get started:
-// https://playwright.dev/docs/intro
-test('visits the app root url', async ({ page }) => {
-  await page.goto('/');
-  await expect(page.locator('div.greetings > h1')).toHaveText('You did it!');
-})
+test('test', async ({ page }) => {
+  await page.goto('http://localhost:5173/');
+});
