+++ src/components/__tests__/UploadImage.test.ts
@@ -1,5 +1,5 @@
 import { it, expect, describe } from 'vitest';
-import useFileSelect from '../fileHandling';
+import useFileSelect from '../useFileUpload';
 
 describe('fileSelect', () => {
   it('should update selectedFile on file input change', async () => {
@@ -17,7 +17,7 @@ describe('fileSelect', () => {
   });
 
   it('should accept only ".jpg,.jpeg,.png" formats', async () => {
-    const file = new File([''], 'document.pdf', { type: 'image/jpg' });
+    const file = new File([''], 'shirt.jpg', { type: 'image/jpg' });
     const event = {
       target: {
         files: [file]
@@ -28,5 +28,19 @@ describe('fileSelect', () => {
 
     handleFileSelect(event);
     expect(selectedFile.value).toBeTruthy();
+  });
+
+  it('should show error when any other formats are uploaded', async () => {
+    const file = new File([''], 'document.pdf', { type: 'application/pdf' });
+    const event = {
+      target: {
+        files: [file]
+      }
+    } as unknown as Event;
+
+    const { handleFileSelect, state } = useFileSelect();
+
+    handleFileSelect(event);
+    expect(state.errorMessage).toMatch(/invalid file type/i);
   });
 });
